% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_decision_table.R
\name{apply_decision_table}
\alias{apply_decision_table}
\title{Apply decision table with stack + pot tracking}
\usage{
apply_decision_table(sim_data, dt_focus, dt_rival = NULL, starting_stack = 100)
}
\arguments{
\item{sim_data}{Simulation dataframe}

\item{dt_focus}{Decision table for focus player}

\item{dt_rival}{Decision table for rival players}

\item{starting_stack}{Starting stack for all players}
}
\value{
Simulation dataframe with betting actions
}
\description{
Apply decision table with stack + pot tracking
}
