% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_5_from_7.R
\name{best_5_from_7}
\alias{best_5_from_7}
\title{Get best 5-card hand from 7 cards}
\usage{
best_5_from_7(cards)
}
\arguments{
\item{cards}{Tibble with 7 cards}
}
\value{
List from evaluate_hand()
}
\description{
Get best 5-card hand from 7 cards
}
