% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_hand.R
\name{best_hand}
\alias{best_hand}
\title{Get best possible 5-card hand from any set of cards}
\usage{
best_hand(cards)
}
\arguments{
\item{cards}{Tibble of 5 to 7 cards}
}
\value{
evaluated hand (list from evaluate_hand)
}
\description{
Get best possible 5-card hand from any set of cards
}
