% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ev.R
\name{calculate_ev}
\alias{calculate_ev}
\title{Calculate EV for each simulated hand}
\usage{
calculate_ev(sim_data)
}
\arguments{
\item{sim_data}{Output from classify_outcomes}
}
\value{
sim_data with EV column
}
\description{
Calculate EV for each simulated hand
}
