% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_outcomes.R
\name{classify_outcomes}
\alias{classify_outcomes}
\title{Classify outcomes based on decisions and actual results}
\usage{
classify_outcomes(sim_data)
}
\arguments{
\item{sim_data}{Output from apply_decision_table}
}
\value{
sim_data with outcome classification
}
\description{
Classify outcomes based on decisions and actual results
}
