% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hands.R
\name{compare_hands}
\alias{compare_hands}
\title{Compare two poker hands}
\usage{
compare_hands(hand1, hand2)
}
\arguments{
\item{hand1}{Tibble of 5 cards}

\item{hand2}{Tibble of 5 cards}
}
\value{
1 if hand1 wins, -1 if hand2 wins, 0 if tie
}
\description{
Compare two poker hands
}
