% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_eval.R
\name{compare_hands_from_eval}
\alias{compare_hands_from_eval}
\title{Compare evaluated hands}
\usage{
compare_hands_from_eval(h1, h2)
}
\arguments{
\item{h1}{evaluated hand}

\item{h2}{evaluated hand}
}
\value{
1, -1, or 0
}
\description{
Compare evaluated hands
}
