% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_cards.R
\name{deal_cards}
\alias{deal_cards}
\title{Deal cards from a deck}
\usage{
deal_cards(deck, n)
}
\arguments{
\item{deck}{A tibble representing a deck of cards}

\item{n}{Number of cards to deal}
}
\value{
A list with:
\describe{
\item{hand}{Tibble of dealt cards}
\item{remaining_deck}{Tibble of remaining cards}
}
}
\description{
Deal cards from a deck
}
