% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_dt_action.R
\name{evaluate_dt_action}
\alias{evaluate_dt_action}
\title{Evaluate DT action for a single player}
\usage{
evaluate_dt_action(row, dt, stage, current_bet, pot, cycle = 1)
}
\arguments{
\item{row}{Single-row dataframe/tibble for player}

\item{dt}{Decision table}

\item{stage}{Current stage}

\item{current_bet}{Current highest bet on street}

\item{pot}{Current pot size}

\item{cycle}{Betting cycle number}
}
\value{
Named list containing:
action,
invest,
matched_rule
}
\description{
Evaluate DT action for a single player
}
