% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_hand.R
\name{evaluate_hand}
\alias{evaluate_hand}
\title{Evaluate a 5-card poker hand}
\usage{
evaluate_hand(cards)
}
\arguments{
\item{cards}{Tibble with columns rank (int) and suit (factor/char)}
}
\value{
List containing rank_class, rank_value, and tiebreaker vector
}
\description{
Evaluate a 5-card poker hand
}
