% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_rule.R
\name{evaluate_rule}
\alias{evaluate_rule}
\title{Evaluate a rule string safely}
\usage{
evaluate_rule(rule, data)
}
\arguments{
\item{rule}{Character string}

\item{data}{Named list (row data)}
}
\value{
TRUE/FALSE
}
\description{
Evaluate a rule string safely
}
