% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_card.R
\name{print_card}
\alias{print_card}
\title{Convert cards to human-readable format}
\usage{
print_card(cards)
}
\arguments{
\item{cards}{A tibble of cards with rank and suit}
}
\value{
A character vector like "A♠", "10♥"
}
\description{
Convert cards to human-readable format
}
