% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_betting_cycle.R
\name{run_betting_cycle}
\alias{run_betting_cycle}
\title{Run one betting cycle for a street}
\usage{
run_betting_cycle(
  sim_data,
  idx,
  active_players,
  dt_focus,
  dt_rival,
  stage,
  cycle,
  current_bet,
  pot,
  street_investments
)
}
\arguments{
\item{sim_data}{Simulation dataframe}

\item{idx}{Row indices for current hand}

\item{active_players}{Logical vector}

\item{dt_focus}{DT for focus player}

\item{dt_rival}{DT for rivals}

\item{stage}{Current street}

\item{cycle}{Betting cycle number}

\item{current_bet}{Current highest bet}

\item{pot}{Current pot}

\item{street_investments}{Numeric vector of
investments already committed this street}
}
\value{
List containing updated:
sim_data,
active_players,
current_bet,
pot,
street_investments,
raise_occurred
}
\description{
Run one betting cycle for a street
}
