% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_many_hands.R
\name{simulate_many_hands}
\alias{simulate_many_hands}
\title{Simulate many Texas Hold'em hands with optional player focus}
\usage{
simulate_many_hands(
  n_sim = 100,
  n_players = 2,
  focus_player = NULL,
  seed = NULL
)
}
\arguments{
\item{n_sim}{Number of simulations}

\item{n_players}{Number of players}

\item{focus_player}{Optional player to track (default = NULL = all players)}

\item{seed}{Optional seed}
}
\value{
Tibble of simulation results
}
\description{
Simulate many Texas Hold'em hands with optional player focus
}
