% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_strength.R
\name{summarise_by_strength}
\alias{summarise_by_strength}
\title{Summarise outcomes by river hand strength}
\usage{
summarise_by_strength(sim_data)
}
\arguments{
\item{sim_data}{Simulation output}
}
\value{
Tibble grouped by hand strength
}
\description{
Summarise outcomes by river hand strength
}
