% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise simulation outcomes}
\usage{
summarise_results(sim_data)
}
\arguments{
\item{sim_data}{Output from simulate_many_hands}
}
\value{
Summary tibble
}
\description{
Summarise simulation outcomes
}
