% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_vs_focushand.R
\name{win_vs_focushand}
\alias{win_vs_focushand}
\title{Win vs Focus Hand Analysis (Decision-Free)}
\usage{
win_vs_focushand(sim_data)
}
\arguments{
\item{sim_data}{Output from simulate_many_hands()}
}
\value{
List of summary tables
}
\description{
Win vs Focus Hand Analysis (Decision-Free)
}
