% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_strategies.R
\docType{data}
\name{dt_tight}
\alias{dt_tight}
\title{Example Poker Strategies}
\format{
A list of tibbles:
\describe{
\item{dt_tight}{Conservative strategy}
\item{dt_base}{Baseline strategy}
\item{dt_loose}{Aggressive strategy}
}
}
\usage{
dt_tight
}
\description{
Predefined decision tables for testing strategy performance.
}
\keyword{datasets}
